replace old trie
